restructure model folder
