--- project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/model/DDStrategy.kt
@@ -1,5 +1,0 @@
-package org.plan.research.minimization.plugin.model
-
-enum class DDStrategy {
-    DD_MIN, PROBABILISTIC_DD
-}
