--- project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/model/HierarchyCollectionStrategy.kt
@@ -1,5 +1,0 @@
-package org.plan.research.minimization.plugin.model
-
-enum class HierarchyCollectionStrategy {
-    FILE_TREE
-}
