+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/model/state/CompilationStrategy.kt
@@ -1,4 +1,4 @@
-package org.plan.research.minimization.plugin.model
+package org.plan.research.minimization.plugin.model.state
 
 enum class CompilationStrategy {
     DUMB, // Return the same exception every time. Used for testing
