+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/model/state/DDStrategy.kt
@@ -1,0 +1,5 @@
+package org.plan.research.minimization.plugin.model.state
+
+enum class DDStrategy {
+    DD_MIN, PROBABILISTIC_DD
+}
