+++ project-minimization-plugin/src/main/kotlin/org/plan/research/minimization/plugin/model/state/HierarchyCollectionStrategy.kt
@@ -1,0 +1,5 @@
+package org.plan.research.minimization.plugin.model.state
+
+enum class HierarchyCollectionStrategy {
+    FILE_TREE
+}
